feat: added contracts and traits for overriding models
- use the interface to typehint the models everywhere it used the explicit class
- updated readme for usage
