+++ src/Contracts/SentEmailUrlClickedModel.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace jdavidbakr\MailTracker\Contracts;
+
+interface SentEmailUrlClickedModel
+{
+    public function getConnectionName();
+    public function email();
+}
