+++ src/Events/EmailSentEvent.php
@@ -2,9 +2,10 @@
 
 namespace jdavidbakr\MailTracker\Events;
 
-use Illuminate\Queue\SerializesModels;
 use Illuminate\Contracts\Queue\ShouldQueue;
-use jdavidbakr\MailTracker\Model\SentEmail;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Queue\SerializesModels;
+use jdavidbakr\MailTracker\Contracts\SentEmailModel;
 
 class EmailSentEvent implements ShouldQueue
 {
@@ -15,10 +16,10 @@ class EmailSentEvent implements ShouldQueue
     /**
      * Create a new event instance.
      *
-     * @param  sent_email  $sent_email
+     * @param  Model|SentEmailModel  $sent_email
      * @return void
      */
-    public function __construct(SentEmail $sent_email)
+    public function __construct(Model|SentEmailModel $sent_email)
     {
         $this->sent_email = $sent_email;
     }
