+++ src/Events/LinkClickedEvent.php
@@ -2,9 +2,10 @@
 
 namespace jdavidbakr\MailTracker\Events;
 
-use Illuminate\Queue\SerializesModels;
 use Illuminate\Contracts\Queue\ShouldQueue;
-use jdavidbakr\MailTracker\Model\SentEmail;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Queue\SerializesModels;
+use jdavidbakr\MailTracker\Contracts\SentEmailModel;
 
 class LinkClickedEvent implements ShouldQueue
 {
@@ -17,11 +18,11 @@ class LinkClickedEvent implements ShouldQueue
     /**
      * Create a new event instance.
      *
-     * @param SentEmail $sent_email
+     * @param Model|SentEmailModel $sent_email
      * @param string $ip_address
      * @param string $link_url
      */
-    public function __construct(SentEmail $sent_email, $ip_address, $link_url)
+    public function __construct(Model|SentEmailModel $sent_email, $ip_address, $link_url)
     {
         $this->sent_email = $sent_email;
         $this->ip_address = $ip_address;
